set Display.errch and use it for errors from mouse/keyboard
i've only seen error EOF from devdraw mouse/keyboard, when i close
a window (at least on macos), or by calling Display.Close().  before
this change, the draw lib would call log.Fatal, ending the entire
program.  now, it will send the error on errch.  it is the caller's
responsibility to pass an errch to Init() that has buffering, if
it wants to be sure it sees the errors.

this doesn't remove the "TODO use errch" comment at Init(), because
not all errors are handled yet. there are still some log.Fatal's,
around allocimage/allocwindow.