Win: Support the errors file
Writing to the errors file is unsupported
and generates unknown file errors.

This change adds a Fid for the errors window to Win
and support for it in fid.
Since errors can only be written to, not read from,
it ensures that the errors Fid is opened in OWRITE mode.