make libdraw (mostly) work on plan 9 under rio
this now mounts $wsys to create a new window of the requested size.
it mounts the rio window at /n/duit.<window-name>, and opens its mouse/keyboard files, writes its label.

bugs: on interrupt (del) of the command, the window isn't removed.

this is loosely based on mischiefs draw9 library.