acme/Watch: use acme log to find out about Put events, instead of kqueue
The acme global log didn't exist when I originally wrote Watch.
Using it instead of kqueue makes Watch portable to non-macOS systems.
It also avoids spurious rebuilds when non-edited files in the directory
change. For example, a problem with the old watch was if you ran
any command that updated a file in the current directory, that update
would trigger it to run again, and on and on.

Now only acme Put operations can trigger a rerun.
(And of course running Get in the /+watch window.)