acme/Watch: more sophisticated command handling
Commands are now treated as rc(1) shell input, confined to a single line.
The prompt shown in the window is now % instead of $.
Any other lines added to the window with a "% " prefix will
be executed the next time the window is refreshed (either by Get
or because a file changed).
Added Quit to get Go stack traces.
Better window handling all around.