plan9/client: refcounted Fsys (#77)
Implement `Fsys.Close` by closing its root `Fid`.

This needed a bunch of reworking to make refcounting of connections
work. We add a new method, `Conn.Release` which allows the
connection to be released when the last `Fid` derived from it
has been closed. The existing `Conn.Close` method retains its
old semantics for backward compatibility (arguably it might be better to change its semantics
and just let callers close the underlying connection directly themselves if they wish).

The closing logic was made more complex by the fact that a `Fid` could
represent both a fid that's been allocated on the server and a
fid that is about to be sent to the server for allocation.
Simplify the invariants by only creating a `Fid` object when we know
that the fid has actually been allocated on the server.

Also guard against concurrent method `Fid` method calls by atomically
closing the fid at exactly the moment we're sure that it will be clunked
(while we hold the write lock, just before sending the RPC request).