acme/Watch: use explicit directory for running command
I've found that it's not uncommon to be running Watch
in a VCS directory, and in the course of rebasing, that
directory is removed and created again. Because the
Watch command is just run in the current directory, that
causes it to fail.

This change uses an explicit directory for the command,
which fixes that failure. A more complete solution would
probably take the directory from the window's tag, but
that would be more involved, because we'd also want
to change the watched directory when that changes,
so leave it for later.