cmd/samterm: Update Mkfifo calls to use unix package
The samterm code calls syscall.Mkfifo in two places.  When compiling on
AIX, this results in a compilation error:

$ go install
./unix.go:25:17: undefined: syscall.Mkfifo
./unix.go:36:22: undefined: syscall.Mkfifo

This is because the Mkfifo call is not implemented for AIX in the syscall
package.  However, Mkfifo is available in the unix package, which is
already imported in cmd/samterm/unix.go.  This change uses the Mkfifo from
unix instead of syscall.

Fixes #123