acmego: fix line deletions at beginning of a file
If the formatter deletes 1 (or more) lines in the beginning of a file,
the corresponding diff output records this with for example "1d0".

Until now, the trailing "0" tripped off the logic and the code just
didn't apply such a deletion. This is now fixed by ignoring the trailing
"0" for line deletion markers.